requirements.txt
opencv_python
numpy
Pillow